Added Gear BaseType.
A Gear object's Base Type is the fundamental type, more specific than the
GearType.  Right now, the GearTypeRunner is used to determine the base type,
and the GearType, if matching, is used as the BaseTYpe.  This will be
expanded, both for the BaseType to be more specific, and to [ultimately] move
determining the BaseType in more appropriate classes.