Allow Recipe to determine RecipeResult name.
Some related recipes produce different results depending on what items are
available and what properties those items have.  For example, a normal, magic,
and rare item of the same base type can give two Orbs of Augmentation if they
are all unidentified, but only one if any are identified.  Thus, the "name' of
the Recipe or RecipeResult can not be determined a priori.  This change adds a
method to the Recipe class, GetResultName, which, if overridden in a child
class, could examine the RecipeResult object to determine the name.