Added ignoring tabs in recipes.
This change is a little bit of a hack.  In order to support ignoring certain
tabs in recipes, I needed to store a boolean value for each tab somewhere.
For convenience, this was added to the Tab model class.  This probably would
be more appropriate in a view model class somewhere.

Also, in order to not simply overwrite any settings related to ignoring tabs
in recipes, when the recipe results is updated, the current view of the stash
is used, not a fresh copy from the cache.