Save tabs ignored in recipes in settings file.
Added a new List to the Settings.xml file, IgnoreTabsInRecipes.  This is
loaded into the Settings.Lists list, and is reference to set whether the
"Ignore in Recipes" menu items are checked.  When this menu item is clicked,
the settings file is updated and saved to reflect the new value.