Added settings to change my leagues and characters.
Added a TreeView in the blank area of the settings tab.  This lists all
characters, grouped by league, with checkboxes for each character and league.
Checkboxes are checked based on whether that item is in the associated
MyCharacters or MyLeagues lists.  Changing a check box value updates the
respective list.

Supporting this required adding an AllCharactersByLeague dictionary to the
ApplicationState class, as well as several properties to the
SettingsViewModel.